requirements.txt
absl==0.0
absl_py==1.4.0
numpy==1.22.4
pandas==1.2.3
PyYAML==6.0.1
scikit_learn==1.3.1
scikit_multiflow==0.5.3
tensorflow==2.4.1
tensorflow_probability==0.12.2
tqdm==4.64.1
wandb==0.15.2
